More robust test timeout
